gateway-service/requirements.txt
fastapi
httpx
requests
python-chess
uvicorn
python-dotenv
asyncpg
PyJWT[crypto]
stripe
pytest
pytest-asyncio
llama-index
llama-index-vector-stores-postgres
pgvector
openai
sentry-sdk[fastapi]
slowapi
redis
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-httpx
opentelemetry-instrumentation-requests
opentelemetry-instrumentation-asyncpg
opentelemetry-instrumentation-redis
opentelemetry-instrumentation-system-metrics
aioboto3
opentelemetry-instrumentation-botocore
